Move shared command helper in cmd/ folder
Since these helpers are only used for commands, we should move it to
where they belong
