+++ cmd/export.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/dcoker/biscuit/cmd/internal/shared"
 	"github.com/dcoker/biscuit/internal/yaml"
-	"github.com/dcoker/biscuit/shared"
 	"github.com/dcoker/biscuit/store"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
