+++ cmd/get.go
@@ -6,8 +6,8 @@ import (
 	"os"
 
 	"github.com/dcoker/biscuit/algorithms"
+	"github.com/dcoker/biscuit/cmd/internal/shared"
 	"github.com/dcoker/biscuit/keymanager"
-	"github.com/dcoker/biscuit/shared"
 	"github.com/dcoker/biscuit/store"
 	"github.com/mattn/go-isatty"
 	"gopkg.in/alecthomas/kingpin.v2"
