+++ cmd/list.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/dcoker/biscuit/shared"
+	"github.com/dcoker/biscuit/cmd/internal/shared"
 	"github.com/dcoker/biscuit/store"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
