+++ cmd/put.go
@@ -12,8 +12,8 @@ import (
 	"sync"
 
 	"github.com/dcoker/biscuit/algorithms"
+	"github.com/dcoker/biscuit/cmd/internal/shared"
 	"github.com/dcoker/biscuit/keymanager"
-	"github.com/dcoker/biscuit/shared"
 	"github.com/dcoker/biscuit/store"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
